24h forecast data(no UI)

+++ app/src/main/java/com/example/weatherapp/response/forecast24h/Temperature.kt
@@ -3,5 +3,5 @@ package com.example.weatherapp.response.forecast24h
 data class Temperature(
     val Unit: String,
     val UnitType: Int,
-    val Value: Int
+    val Value: Double
 )
